:dizzy:[EDIT] main page 뷰 전체적인 수정
