+++ src/components/mainPage/MainPage.tsx
@@ -4,16 +4,12 @@ import { useEffect, useState, lazy, Suspense } from "react"
 import { Provider } from "react-redux"
 import store from "@/redux/store/store"
 import { useRouter } from "next/navigation"
-import CircularUnderLoad from "../CircularUnderLoad"
 import dynamic from "next/dynamic"
+import { getCookie } from "@/util/async/Cookie"
 
-// const MainBtnTable = lazy(() => import("./MainBtnTable"))
-// const MainTable = lazy(() => import("./MainTable"))
-const MainBtnTable = dynamic(() => import("./MainBtnTable"), { ssr: false })
 const MainTable = dynamic(() => import("./MainTable"), { ssr: false })
 
 export default function MainPageComponent() {
-  const [summary, setSummary] = useState(false)
   const router = useRouter()
 
   const prefetchOptions = {
@@ -24,30 +20,16 @@ export default function MainPageComponent() {
     router.prefetch("/main")
   }, [router])
 
+  useEffect(() => {
+    const cookie = getCookie("access_token")
+    if (!cookie) {
+      router.push("/login")
+    }
+  }, [])
   return (
     <div id="root">
       <Provider store={store}>
-        {summary ? (
-          <Suspense
-            fallback={
-              <div>
-                <CircularUnderLoad />
-              </div>
-            }
-          >
-            <MainTable setSummary={setSummary} />
-          </Suspense>
-        ) : (
-          <Suspense
-            fallback={
-              <div>
-                <CircularUnderLoad />
-              </div>
-            }
-          >
-            <MainBtnTable setSummary={setSummary} />
-          </Suspense>
-        )}
+        <MainTable />
       </Provider>
     </div>
   )
